+++ src/XUnity.AutoTranslator.Plugin.Core/Extensions/StringBuilderExtensions.cs
@@ -20,5 +20,23 @@ namespace XUnity.AutoTranslator.Plugin.Core.Extensions
          var lastChar = builder[ builder.Length - 1 ];
          return char.IsWhiteSpace( lastChar ) || lastChar == '\n';
       }
+
+      internal static StringBuilder Reverse( this StringBuilder text )
+      {
+         if( text.Length > 1 )
+         {
+            int pivotPos = text.Length / 2;
+            for( int i = 0; i < pivotPos; i++ )
+            {
+               int iRight = text.Length - ( i + 1 );
+               char rightChar = text[ i ];
+               char leftChar = text[ iRight ];
+               text[ i ] = leftChar;
+               text[ iRight ] = rightChar;
+            }
+         }
+
+         return text;
+      }
    }
 }