+++ src/XUnity.AutoTranslator.Plugin.Core/WhitespaceHandlingStrategy.cs
@@ -8,6 +8,6 @@ namespace XUnity.AutoTranslator.Plugin.Core
    internal enum WhitespaceHandlingStrategy
    {
       TrimPerNewline,
-      AllOccurrences
+      None
    }
 }